backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Keval Shah <keval.arhan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.32.1"
fastapi = "^0.115.5"
sqlalchemy = "^2.0.36"
sqlmodel = "^0.0.22"
asyncpg = "^0.30.0"
psycopg2 = "^2.9.10"
greenlet = "^3.1.1"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.6.1"
python-jose = "^3.3.0"
httpx = "^0.28.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "app.main:start"

